refactor(back/users): :recycle: changed activate to other class and user auth routes
