+++ backend/src/test/java/com/tfg/cultura/api/users/controller/UserAuthControllerTest.java
@@ -10,8 +10,6 @@ import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.DisabledException;
 import org.springframework.test.web.servlet.MockMvc;
 
-import com.tfg.cultura.api.core.exception.UnathenticatedException;
-import com.tfg.cultura.api.users.exception.SelfActivationNotAllowedException;
 import com.tfg.cultura.api.users.exception.UserAlreadyExistsException;
 import com.tfg.cultura.api.users.exception.UserNotFoundException;
 import com.tfg.cultura.api.users.exception.UsersExceptionHandler;
@@ -34,11 +32,9 @@ class UserAuthControllerTest extends BaseControllerTest {
     @Mock
     private UserAuthService userService;
 
-    private static final String BASE_URL = "/api/users";
+    private static final String BASE_URL = "/api/users/auth";
     private static final String REGISTER_URL = BASE_URL + "/register";
     private static final String LOGIN_URL = BASE_URL + "/login";
-    private static final String USER_URL = BASE_URL + "/{id}";
-    private static final String ACTIVATE_URL = USER_URL + "/activate";
 
     private UserRegisterRequest registerRequest;
     private UserLoginRequest loginRequest;
@@ -165,52 +161,6 @@ class UserAuthControllerTest extends BaseControllerTest {
                 "test.pdf",
                 "application/pdf",
                 "dummy pdf content".getBytes());
-    }
-
-    // ================ ACTIVATE USER ================
-
-    @Test
-    void activate_user_sucess() throws Exception {
-        when(userService.activateUser(any())).thenReturn(userResponse);
-
-        mockMvc.perform(put(ACTIVATE_URL, "123"))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.username").value(userResponse.getUsername()));
-    }
-
-    @Test
-    void activate_user_fail_unexisting_user() throws Exception {
-        String message = "El usuario con id 123 no existe";
-        UserNotFoundException ex = new UserNotFoundException(message);
-        when(userService.activateUser(any())).thenThrow(ex);
-
-        mockMvc.perform(put(ACTIVATE_URL, "123"))
-                .andExpect(status().isNotFound())
-                .andExpect(jsonPath("$.message").value(message));
-    }
-
-    @Test
-    void activate_user_fail_self_activation() throws Exception {
-        String userId = "123";
-        String message = String.format("El usuario %s con id %s ha intentado activar su propio usuario",
-                userResponse.getUsername(), userId);
-        SelfActivationNotAllowedException ex = new SelfActivationNotAllowedException(message);
-        when(userService.activateUser(userId)).thenThrow(ex);
-
-        mockMvc.perform(put(ACTIVATE_URL, userId))
-                .andExpect(status().isForbidden())
-                .andExpect(jsonPath("$.message").value(message));
-    }
-
-    @Test
-    void activate_user_fail_unathenticated() throws Exception {
-        String message = "No se ha podido obtener la autenticación del usuario";
-        UnathenticatedException ex = new UnathenticatedException(message);
-        when(userService.activateUser(any())).thenThrow(ex);
-
-        mockMvc.perform(put(ACTIVATE_URL, "123"))
-                .andExpect(status().isUnauthorized())
-                .andExpect(jsonPath("$.message").value(message));
     }
 
 }
